+++ tests/symbolics/test_sympy_backend.py
@@ -11,7 +11,7 @@ Tests for the SympyExpression implementation.
 """
 
 import pytest
-from sympy import sympify
+from sympy import E, cos, exp, pi, sin, sqrt, sympify
 
 from bartiq.errors import BartiqCompilationError
 from bartiq.symbolics import sympy_backend
@@ -33,6 +33,33 @@ from bartiq.symbolics import sympy_backend
 )
 def test_is_sympy_int(expression, expected):
     assert sympy_backend.is_constant_int(expression) == expected
+
+
+@pytest.mark.parametrize(
+    "expression_str, expected",
+    [
+        # Known constants
+        ("pi", pi),
+        ("PI", pi),
+        ("e", E),
+        ("E", E),
+        # Expressions with constants
+        ("2 * pi", 2 * pi),
+        ("e + 1", E + 1),
+        ("sqrt(-cos(3*pi))", sqrt(-cos(3 * pi))),
+        # Mixed case
+        ("2 * PI", 2 * pi),
+        ("sin(-0.25*e)", sin(-0.25 * E)),
+        ("-exp(Pi / 4)", -exp(pi / 4)),
+        # No constants
+        ("x", sympify("x")),
+        ("x + y", sympify("x + y")),
+    ],
+)
+def test_parse_constant(expression_str, expected):
+    expr = sympy_backend.as_expression(expression_str)
+    expr = sympy_backend.parse_constant(expr)
+    assert expr == expected
 
 
 def test_value_of_returns_none_if_numerical_evaluation_is_not_possible():