+++ tests/symbolics/test_sympy_interpreter.py
@@ -50,7 +50,6 @@ from sympy import (
     frac,
     im,
     log,
-    multiplicity,
     prod,
     re,
     sec,
@@ -64,7 +63,9 @@ from sympy import (
 from sympy.codegen.cfunctions import exp2, log2, log10
 from sympy.core.numbers import S as sympy_constants
 
-from bartiq.symbolics.sympy_interpreter import SPECIAL_PARAMS, Round, parse_to_sympy
+from bartiq.symbolics.sympy_backends import parse_to_sympy
+from bartiq.symbolics.sympy_interpreter import SPECIAL_PARAMS, Round, multiplicity
+from bartiq.symbolics.sympy_interpreter import parse_to_sympy as legacy_parse_to_sympy
 from bartiq.symbolics.sympy_serializer import serialize_expression
 
 
@@ -244,9 +245,9 @@ PARSE_TEST_CASES = [
     ("PI ^ 2", Pi**2),
     ("round(PI ^ 2)", 10),
     ("6.02E23 * 8.048", 4.844896e24),
-    ("sin(PI / 2)", 1.0),
-    ("10 + sin(PI / 4) ^ 2", 10.5),
-    ("exp(0)", 1.0),
+    ("sin(PI / 2)", 1),
+    ("10 + sin(PI / 4) ^ 2", Rational(21, 2)),
+    ("exp(0)", 1),
     ("exp(1)", E),
     ("2 ^ 3 ^ 2", 512),
     ("(2 ^ 3) ^ 2", 64),
@@ -363,24 +364,30 @@ PARSE_TEST_CASES = [
     ("x / y // z", (x / y) // z),
     ("x ^ y ** z", x ** (y**z)),
     # Special functions
-    ("multiplicity(x, y)", Function("multiplicity")(x, y)),
-    ("multiplicity(x, 2)", Function("multiplicity")(x, 2)),
-    ("multiplicity(2, y)", Function("multiplicity")(2, y)),
+    ("multiplicity(x, y)", multiplicity(x, y)),
+    ("multiplicity(x, 2)", multiplicity(x, 2)),
+    ("multiplicity(2, y)", multiplicity(2, y)),
     ("multiplicity(2, 40)", multiplicity(2, 40)),
     # Expressions with wildcard
     ("sum(~.X)", Function("sum")(Symbol("~.X"))),
     ("max(~.X)", Function("max")(Symbol("~.X"))),
     ("min(~.X)", Function("min")(Symbol("~.X"))),
+    # Expressions containing "in" keyword
+    ("in", Symbol("in")),
+    ("a.in", Symbol("a.in")),
+    ("a.in.b", Symbol("a.in.b")),
+    ("in.b", Symbol("in.b")),
 ]
 
 
 @pytest.mark.parametrize("expression, expected_sympy_expression", PARSE_TEST_CASES)
-def test_parse_to_sympy(expression, expected_sympy_expression):
+@pytest.mark.parametrize("parse", [parse_to_sympy, legacy_parse_to_sympy])
+def test_parse_to_sympy(expression, expected_sympy_expression, parse):
     """Tests for the sympy expression parser."""
-    sympy_expression = parse_to_sympy(expression, debug=True)
+    sympy_expression = parse(expression, debug=True)
     assert sympy_expression == expected_sympy_expression
 
     # Test round-trip via cast to string
     new_expression = serialize_expression(sympy_expression)
-    sympy_expression = parse_to_sympy(new_expression)
+    sympy_expression = parse(new_expression)
     assert sympy_expression == expected_sympy_expression